move countries and region loading from occ user payment adapter to occ site adapter
