+++ projects/core/src/occ/adapters/user/occ-user-payment.adapter.ts
@@ -1,6 +1,6 @@
 import { Injectable } from '@angular/core';
 import { UserPaymentAdapter } from '../../../user/connectors/payment/user-payment.adapter';
-import { HttpClient, HttpHeaders, HttpParams } from '@angular/common/http';
+import { HttpClient, HttpHeaders } from '@angular/common/http';
 import { OccEndpointsService } from '../../services/occ-endpoints.service';
 import { Observable, throwError } from 'rxjs';
 import { catchError, map } from 'rxjs/operators';
@@ -8,18 +8,9 @@ import { Occ } from '../../occ-models/occ.models';
 import { PaymentDetails } from '../../../model/cart.model';
 import { ConverterService } from '../../../util/converter.service';
 import { PAYMENT_DETAILS_NORMALIZER } from '../../../cart/connectors/payment/converters';
-import { Country, Region } from '../../../model/address.model';
-import {
-  COUNTRY_NORMALIZER,
-  REGION_NORMALIZER,
-} from '../../../user/connectors/payment/converters';
 
 const USER_ENDPOINT = 'users/';
 const PAYMENT_DETAILS_ENDPOINT = '/paymentdetails';
-const COUNTRIES_ENDPOINT = 'countries';
-const REGIONS_ENDPOINT = 'regions';
-const COUNTRIES_TYPE_BILLING = 'BILLING';
-const COUNTRIES_TYPE_SHIPPING = 'SHIPPING';
 
 @Injectable()
 export class OccUserPaymentAdapter implements UserPaymentAdapter {
@@ -72,40 +63,5 @@ export class OccUserPaymentAdapter implements UserPaymentAdapter {
         { headers }
       )
       .pipe(catchError((error: any) => throwError(error)));
-  }
-
-  loadBillingCountries(): Observable<Country[]> {
-    return this.http
-      .get<Occ.CountryList>(this.occEndpoints.getEndpoint(COUNTRIES_ENDPOINT), {
-        params: new HttpParams().set('type', COUNTRIES_TYPE_BILLING),
-      })
-      .pipe(
-        catchError((error: any) => throwError(error.json())),
-        map(countryList => countryList.countries),
-        this.converter.pipeableMany(COUNTRY_NORMALIZER)
-      );
-  }
-
-  loadDeliveryCountries(): Observable<Country[]> {
-    return this.http
-      .get<Occ.CountryList>(this.occEndpoints.getEndpoint(COUNTRIES_ENDPOINT), {
-        params: new HttpParams().set('type', COUNTRIES_TYPE_SHIPPING),
-      })
-      .pipe(
-        catchError((error: any) => throwError(error.json())),
-        map(countryList => countryList.countries),
-        this.converter.pipeableMany(COUNTRY_NORMALIZER)
-      );
-  }
-
-  loadRegions(countryIsoCode: string): Observable<Region[]> {
-    const regionsEndpoint = `${COUNTRIES_ENDPOINT}/${countryIsoCode}/${REGIONS_ENDPOINT}`;
-    return this.http
-      .get<Occ.RegionList>(this.occEndpoints.getEndpoint(regionsEndpoint))
-      .pipe(
-        catchError((error: any) => throwError(error.json())),
-        map(regionList => regionList.regions),
-        this.converter.pipeableMany(REGION_NORMALIZER)
-      );
   }
 }
